feat: 비디오 상세 페이지로 이동시 정보 쿠키 저장 기능 추가 #179
