+++ src/utils/cookieServer.ts
@@ -1,7 +1,15 @@
+// 서버 함수
 import { cookies } from 'next/headers'
 
-// 서버 함수
-export const getVideoInfoToCookie = () => {
+export interface IVideoInfoToCookie {
+  channelTitle: string
+  videoId: string
+  channelId: string
+  title: string
+  thumbnailsUrl: string
+}
+
+export const getVideoInfoToCookie = (): IVideoInfoToCookie | null => {
   const cookieStore = cookies()
   const detailVideoInfo = cookieStore.get('detail-video-info')
   if (detailVideoInfo) {
